Fix general

+++ config/db.js
@@ -9,7 +9,7 @@ const sequelize = new Sequelize(
     dialect: process.env.DB_DIALECT,
     logging: console.log,
     logging: function (str) {
-        console.log(str);
+      console.log(str);
     }
   },
 );
@@ -17,7 +17,13 @@ const sequelize = new Sequelize(
 async function authenticate() {
   try {
     await sequelize.authenticate();
-    await sequelize.sync({ force: true});
+    console.log(process.env.DB_NAME, process.env.DB_FORCE);
+
+    if (process.env.DB_FORCE === 'true') {
+      console.log('Recreación de base de datos ...');
+      await sequelize.sync({ force: true });
+    };
+
     console.log("Conexión a la base de datos satisfactoriamente.");
     //const associations = require("./../api/models/associations/core");
